Prevent merchant bulk-buy cost overflow from granting free items
CanAfford/SpendCosts/RefundCosts multiplied unit cost by quantity in int,
so a 50x buy of a high-cost stock row could wrap negative. SpendGold
treats amount <= 0 as success, which made the purchase free. Use checked
long multiplication and reject overflow as unaffordable; format bulk
labels with long as well.

+++ Assets/6.Scripts/NPC/ShopCostFormatter.cs
@@ -14,7 +14,7 @@ public static class ShopCostFormatter
         if (!showBulk)
             return $"{amount}g";
 
-        int bulk = amount * BulkPurchaseQuantity;
+        long bulk = (long)amount * BulkPurchaseQuantity;
         return $"{amount}g ({bulk}g x{BulkPurchaseQuantity})";
     }
 
@@ -26,7 +26,7 @@ public static class ShopCostFormatter
         if (!showBulk)
             return $"• {amount} Gold";
 
-        int bulk = amount * BulkPurchaseQuantity;
+        long bulk = (long)amount * BulkPurchaseQuantity;
         return $"• {amount} Gold ({bulk} Gold x{BulkPurchaseQuantity})";
     }
 
@@ -38,7 +38,7 @@ public static class ShopCostFormatter
         if (!showBulk)
             return $"{amount} {itemName}";
 
-        int bulk = amount * BulkPurchaseQuantity;
+        long bulk = (long)amount * BulkPurchaseQuantity;
         return $"{amount} {itemName} ({bulk} {itemName} x{BulkPurchaseQuantity})";
     }
